Standardised file names to snake case and C++ headers to end in .hpp. Taking advantage of the major version change we need for adding the MKS controller implementation (which to avoid confusion includes renaming the original StepperController class)
