+++ include/umrt-arm-firmware-lib/utils.hpp
@@ -3,8 +3,8 @@
  * A collection of helper functions related to encoding/decoding data for communication over a Firmata link.
  */
 
-#ifndef UMRT_ARM_FIRMWARE_LIB_UTILS_H
-#define UMRT_ARM_FIRMWARE_LIB_UTILS_H
+#ifndef UMRT_ARM_FIRMWARE_LIB_UTILS_HPP
+#define UMRT_ARM_FIRMWARE_LIB_UTILS_HPP
 
 #include <cstdint>
 #include <string>
@@ -193,4 +193,4 @@ inline std::string decode_string(const std::vector<uint8_t>& data) {
     return { data.cbegin(), data.cend() };
 }
 
-#endif //UMRT_ARM_FIRMWARE_LIB_UTILS_H
+#endif //UMRT_ARM_FIRMWARE_LIB_UTILS_HPP
